+++ podcasts/Onboarding/Plus/NewUpgrade/UpgradeAccountView.swift
@@ -13,6 +13,7 @@ struct UpgradeAccountView: View {
     @State private var flash: Bool = false
 
     enum ScrollPosition: String {
+        case firstPage
         case secondPage
     }
 
@@ -20,44 +21,7 @@ struct UpgradeAccountView: View {
         VStack(alignment: .leading, spacing: 0) {
             header
             Spacer().frame(height: 24)
-            ScrollViewReader { proxy in
-                ScrollView {
-                    VStack(alignment: .leading, spacing: 24) {
-                        title
-                        UpgradeFeaturesView(features: model.features)
-                        if model.isFreeTrialAvailable {
-                            Button {
-                                expand = true
-                                withAnimation {
-                                    proxy.scrollTo(ScrollPosition.secondPage, anchor: .bottom)
-                                }
-                            } label: {
-                                Text(L10n.subscriptionPlanFreeTrialInfoLink)
-                                    .font(.subheadline)
-                                    .foregroundColor(theme.primaryInteractive01)
-                            }
-                        }
-                        Spacer()
-                        if expand, model.isFreeTrialAvailable {
-                            VStack {
-                                HStack {
-                                    Spacer()
-                                    Text("Placeholder View - Variant B")
-                                        .foregroundStyle(theme.primaryText01)
-                                    Spacer()
-                                }
-                            }
-                            .padding(.vertical, 200)
-                            .id(ScrollPosition.secondPage)
-                        } else {
-                            EmptyView()
-                                .id(ScrollPosition.secondPage)
-                        }
-                    }
-                }
-                .scrollIndicators(.visible)
-                .withScrollFlashIndicator(trigger: flash)
-            }
+            scrollableContent
             UpgradeProductsView(model: model)
         }
         .padding(.horizontal, 24)
@@ -91,6 +55,54 @@ struct UpgradeAccountView: View {
         }
     }
 
+    var scrollableContent: some View {
+        GeometryReader() { sizeProxy in
+            ScrollViewReader { proxy in
+                ScrollView {
+                    VStack(alignment: .leading, spacing: 24) {
+                        title
+                            .id(ScrollPosition.firstPage)
+                        UpgradeFeaturesView(features: model.features)
+                        if model.isFreeTrialAvailable {
+                            Button {
+                                expand = true
+                                withAnimation {
+                                    proxy.scrollTo(ScrollPosition.secondPage, anchor: .top)
+                                }
+                            } label: {
+                                Text(L10n.subscriptionPlanFreeTrialInfoLink)
+                                    .font(size: 15, style: .subheadline, weight: .medium)
+                                    .foregroundColor(theme.primaryInteractive01)
+                            }
+                        }
+                        if expand, model.isFreeTrialAvailable {
+                            UpgradeTimelineView(events: model.timelineEvents)
+                                .id(ScrollPosition.secondPage)
+                                .frame(minHeight: sizeProxy.size.height)
+                        } else {
+                            Spacer()
+                                .id(ScrollPosition.secondPage)
+                                .frame(height: 8)
+                        }
+                    }
+                }
+                .scrollIndicators(.visible)
+                .withScrollFlashIndicator(trigger: flash)
+                .overlay(alignment: .bottom, content: {
+                    gradientSpacer
+                })
+                .onChange(of: model.selectedProduct) { _ in
+                    if !model.isFreeTrialAvailable {
+                        withAnimation {
+                            expand = false
+                            proxy.scrollTo(ScrollPosition.firstPage)
+                        }
+                    }
+                }
+            }
+        }
+    }
+
     var title: some View {
         HStack {
             Text(model.upgradeTier.header)
@@ -99,6 +111,18 @@ struct UpgradeAccountView: View {
                 .foregroundColor(theme.primaryText01)
             Spacer()
         }
+    }
+
+    var gradientSpacer: some View {
+        HStack() {
+            Spacer()
+        }
+        .frame(height: 40)
+        .background(LinearGradient(colors: [
+            theme.primaryUi01.opacity(0),
+            theme.primaryUi01.opacity(1)
+        ], startPoint: UnitPoint.top, endPoint: UnitPoint.bottom))
+        .allowsHitTesting(false)
     }
 }